Backtest refactor
- Preparing strategy class to multiprocessing simulation
- Adding status expired and removing order size when running backtests
- Summarizer class moved to backtester module
- Cerebro does not store backtests results. Now finish function within Strategy class saves a small csv at the end of the backtest. This way, it does not consume extra RAM when running optstrategy

+++ src/oandatradingbot/utils/messages.py
@@ -38,7 +38,7 @@ class Messages:
         return response
 
     def buy_order_placed(
-        self, size: int, pair: str, price: float, id=None
+        self, size: float, pair: str, price: float, id=None
     ) -> str:
         if self.lang == "ES-ES":
             response = (
@@ -137,7 +137,7 @@ class Messages:
         return response
 
     def sell_order_placed(
-        self, size: int, pair: str, price: float, id=None
+        self, size: float, pair: str, price: float, id=None
     ) -> str:
         if self.lang == "ES-ES":
             response = (
